requirements.txt
Flask==3.1.0
flask-cors==5.0.0
Flask-Injector==0.15.0
gunicorn==23.0.0
protobuf==5.29.3
aiohttp==3.11.12
wheel==0.44.0
click==8.1.8
MarkupSafe==3.0.2
Werkzeug==3.1.3
mkdocstrings[python]==0.28.1
mkdocs-material==9.6.4
itsdangerous==2.2.0
Jinja2==3.1.5
mkdocs==1.6.1
SQLAlchemy==2.0.38
tabulate==0.9.0
typing-inspect==0.9.0
typing_extensions==4.12.2
cryptography==44.0.1
pika==1.3.2
sql==0.4.0
psycopg2-binary==2.9.7
yt-dlp==2024.7.16

# Utils
transformers==4.45.2
tiktoken==0.9.0
python-dotenv==1.0.1
pydantic==2.10.6
flask-executor==1.0.0
flask-sqlalchemy==3.0.0
retry==0.9.2
XlsxWriter==3.2.2
PyMuPDF==1.23.6
pytesseract==0.3.10
openpyxl==3.1.2
networkx==3.2.1
python-pptx==0.6.23
unstructured==0.15.12
GitPython==3.1.40
beautifulsoup4==4.12.2
sentry-sdk==2.21.0
pdfplumber==0.11.0
aio-pika==9.0.5

# AI & core services
openai==1.63.0
langchain==0.3.18
langchainhub==0.1.21
langchain_openai==0.3.5
langchain_qdrant==0.2.0
langchain_neo4j==0.4.0
langchain_community==0.3.16
langgraph==0.2.69


# Data
qdrant-client==1.13.2


# File processing libraries for chat uploads
PyMuPDF==1.23.6
pdfplumber==0.11.0
pytesseract==0.3.10 # image OCR
python-docx==0.8.11 # DOCX files
openpyxl==3.1.2 # excel
python-pptx==0.6.23 # PowerPoint
beautifulsoup4==4.12.2 # HTML parsing
pysrt==1.1.2 # SRT subtitle files
Pillow==10.0.0 # Image processing for OCR
docx2txt==0.8 # Alternative DOCX processing
# added
azure-storage-blob
pyodbc

